auto comment

+++ common.h
@@ -2,7 +2,7 @@
 #define dt 1000000
 
 extern int n;
-extern long  double *x, *y, *z, *new_x, *new_y, *new_z, *new_sx, *new_sy, *new_sz, *mass, *sx, *sy, *sz;
+extern double *x, *y, *z, *new_x, *new_y, *new_z, *new_sx, *new_sy, *new_sz, *mass, *sx, *sy, *sz;
 
 void print_csv_body(int i);
 void print_csv_bodies();
