+++ sequential-barneshut.h
@@ -1,12 +1,12 @@
 extern int n;
-extern long double *x, *y, *z, *mass, *sx, *sy, *sz;
+extern double *x, *y, *z, *mass, *sx, *sy, *sz;
 
 typedef struct bnode{
         int body; // >= 0 se contiene un corpo, -1 se contiene zero corpi, -2 se contiene più di un corpo
         int depth;
-        long int max_x, max_y, max_z, min_x, min_y, min_z;
-        long double x, y, z;
-        long double mass; 
+        int max_x, max_y, max_z, min_x, min_y, min_z;
+        double x, y, z;
+        double mass; 
         struct bnode *o0, *o1, *o2, *o3, *o4, *o5, *o6, *o7;
 } bnode;
 
@@ -30,12 +30,12 @@ void print_tree(bnode* node);
 
 void print_node(bnode* node);
 
-long int get_bound();
+int get_bound();
 void build_barnes_tree(bnode* root);
 void destroy_barnes_tree(bnode* root);
 void generate_empty_children(bnode* node);
-bnode* get_octant(bnode* node, long double x, long double y, long double z);
-void update(bnode* node, int body, long double x, long double y, long double z, long double mass);
+bnode* get_octant(bnode* node, double x, double y, double z);
+void update(bnode* node, int body, double x, double y, double z, double mass);
 void insert_body(bnode* node, int body);
 void compute_barnes_forces(bnode* node, int body, double theta);
 void compute_barnes_forces_all(bnode* root, double theta);
